backend/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
psycopg==3.2.3
psycopg-binary==3.2.3
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2024.2
sqlparse==0.5.3
tzdata==2024.2

